# docs/api/openapi.v0.1.yaml
openapi: 3.1.0
info:
  title: Knowledge Base Agent Assistant API
  version: 0.1.0
  description: v0.1 frontend-backend API contract for the internal Knowledge Base Agent Assistant.
servers:
  - url: http://localhost:8000/api/v1
    description: Local backend
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Error"
  /users/{user_id}:
    patch:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /users/{user_id}/disable:
    post:
      tags: [Users]
      summary: Disable user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Disabled user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{user_id}/enable:
    post:
      tags: [Users]
      summary: Enable user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Enabled user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{user_id}/reset-password:
    post:
      tags: [Users]
      summary: Reset user password
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResponse"
  /knowledge-bases:
    get:
      tags: [KnowledgeBases]
      summary: List knowledge bases
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/KnowledgeBaseStatus"
      responses:
        "200":
          description: Knowledge base list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseListResponse"
    post:
      tags: [KnowledgeBases]
      summary: Create knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseCreateRequest"
      responses:
        "201":
          description: Created knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
  /knowledge-bases/{kb_id}:
    get:
      tags: [KnowledgeBases]
      summary: Get knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
    patch:
      tags: [KnowledgeBases]
      summary: Update knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseUpdateRequest"
      responses:
        "200":
          description: Updated knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
    delete:
      tags: [KnowledgeBases]
      summary: Soft delete knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "204":
          description: Deleted
...
  /knowledge-graph:
    get:
      tags: [KnowledgeGraph]
      summary: Get document summary relationship graph
      parameters:
        - name: knowledge_base_id
          in: query
          schema:
            type: string
        - name: include_cross_knowledge_base
          in: query
          schema:
            type: boolean
            default: true
        - name: min_similarity
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0.45
      responses:
        "200":
          description: Knowledge graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraph"
  /knowledge-graph/refresh:
    post:
      tags: [KnowledgeGraph]
      summary: Rebuild document relationships and community summaries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force_embeddings:
                  type: boolean
                  default: false
      responses:
        "202":
          description: Rebuild requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraph"
  /knowledge-bases/{knowledge_base_id}/community-summary:
    get:
      tags: [KnowledgeGraph]
      summary: Get knowledge-base community summary
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Community summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunitySummary"
  /knowledge-bases/{knowledge_base_id}/retrieval/search:
    post:
      tags: [Retrieval]
      summary: Search chunks within one knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrievalSearchRequest"
      responses:
        "200":
          description: Retrieval results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrievalSearchResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
    KnowledgeBaseId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ISODateTime:
      type: string
      format: date-time
    UserRole:
      type: string
      enum: [admin, user]
    KnowledgeBaseStatus:
      type: string
      enum: [active, deleting, deleted]
...
    ErrorEnvelope:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorEnvelope"
...
    User:
      type: object
      required: [id, username, display_name, role, is_active, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        is_active:
          type: boolean
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        last_login_at:
          anyOf:
            - $ref: "#/components/schemas/ISODateTime"
            - type: "null"
    UserCreateRequest:
      type: object
      required: [email, username, display_name, password, role]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type: string
        password:
          type: string
          format: password
        role:
          $ref: "#/components/schemas/UserRole"
    UserUpdateRequest:
      type: object
      properties:
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
    ResetPasswordRequest:
      type: object
      required: [new_password]
      properties:
        new_password:
          type: string
          format: password
    ResetPasswordResponse:
      type: object
      required: [user_id, reset_at]
      properties:
        user_id:
          type: string
        reset_at:
          $ref: "#/components/schemas/ISODateTime"
    UserListResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/User"
    PaginatedResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    KnowledgeBase:
      type: object
      required: [id, name, status, file_count, chunk_count, created_by, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/KnowledgeBaseStatus"
        file_count:
          type: integer
        chunk_count:
          type: integer
        created_by:
          type: string
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        updated_at:
          $ref: "#/components/schemas/ISODateTime"
    KnowledgeBaseCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    KnowledgeBaseUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/KnowledgeBaseStatus"
    KnowledgeBaseListResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/KnowledgeBase"
...
    KnowledgeGraphNode:
      type: object
      required:
        - id
        - file_id
        - document_summary_id
        - file_name
        - file_ext
        - knowledge_base_id
        - knowledge_base_name
        - summary
        - summary_status
        - relation_count
      properties:
        id:
          type: string
        file_id:
          type: string
        document_summary_id:
          type: string
        file_name:
          type: string
        file_ext:
          type: string
        knowledge_base_id:
          type: string
        knowledge_base_name:
          type: string
        summary:
          type: string
        summary_status:
          type: string
        relation_count:
          type: integer
    KnowledgeGraphEdge:
      type: object
      required: [id, source, target, similarity, cross_knowledge_base]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        similarity:
          type: number
          minimum: 0
          maximum: 1
        cross_knowledge_base:
          type: boolean
    CommunitySummary:
      type: object
      required:
        - knowledge_base_id
        - knowledge_base_name
        - status
        - document_count
        - prompt_version
        - reduction_level
      properties:
        knowledge_base_id:
          type: string
        knowledge_base_name:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, not_ready]
        summary:
          type:
            - string
            - "null"
        document_count:
          type: integer
        model_name:
          type:
            - string
            - "null"
        prompt_version:
          type: string
        reduction_level:
          type: integer
        error_code:
          type:
            - string
            - "null"
        error_message:
          type:
            - string
            - "null"
        updated_at:
          anyOf:
            - $ref: "#/components/schemas/ISODateTime"
            - type: "null"
    KnowledgeGraph:
      type: object
      required:
        - status
        - document_count
        - total_document_count
        - summarized_document_count
        - pending_summary_count
        - failed_summary_count
        - not_ready_document_count
        - relation_count
        - similarity_threshold
        - max_relations_per_document
        - nodes
        - edges
        - communities
      properties:
        status:
          type: string
          enum: [pending, running, completed, failed]
        source_fingerprint:
          type:
            - string
            - "null"
        document_count:
          type: integer
        total_document_count:
          type: integer
        summarized_document_count:
          type: integer
        pending_summary_count:
          type: integer
        failed_summary_count:
          type: integer
        not_ready_document_count:
          type: integer
        relation_count:
          type: integer
        embedding_model:
          type:
            - string
            - "null"
        similarity_threshold:
          type: number
        max_relations_per_document:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeGraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeGraphEdge"
        communities:
          type: array
          items:
            $ref: "#/components/schemas/CommunitySummary"
...
    RetrievalSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        vector_top_k:
          type: integer
          minimum: 1
          maximum: 100
          default: 30
        full_text_top_k:
          type: integer
          minimum: 1
          maximum: 100
          default: 30
        top_k:
          type: integer
          minimum: 1
          maximum: 20
          default: 8
    RetrievalResultItem:
      type: object
      required: [chunk_id, file_id, file_name, source_locator, excerpt, score, source]
      properties:
        chunk_id:
          type: string
        file_id:
          type: string
        file_name:
          type: string
        source_locator:
          type: string
        excerpt:
          type: string
        score:
          type: number
        source:
          type: string
          enum: [vector, full_text, hybrid]
    RetrievalSearchResponse:
      type: object
      required: [knowledge_base_id, query, items, total]
      properties:
        knowledge_base_id:
          type: string
        query:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalResultItem"
        total:
          type: integer
...